+++ imblearn/under_sampling/_prototype_selection/_edited_nearest_neighbours.py
@@ -10,12 +10,12 @@ import numbers
 from collections import Counter
 
 import numpy as np
+from scipy.stats import mode
 from sklearn.utils import _safe_indexing
 from sklearn.utils._param_validation import HasMethods, Interval, StrOptions
 
 from ...utils import Substitution, check_neighbors_object
 from ...utils._docstring import _n_jobs_docstring
-from ...utils.fixes import _mode
 from ..base import BaseCleaningSampler
 
 SEL_KIND = ("all", "mode")
@@ -168,7 +168,7 @@ class EditedNearestNeighbours(BaseCleaningSampler):
                 nnhood_idx = self.nn_.kneighbors(X_class, return_distance=False)[:, 1:]
                 nnhood_label = y[nnhood_idx]
                 if self.kind_sel == "mode":
-                    nnhood_label, _ = _mode(nnhood_label, axis=1)
+                    nnhood_label, _ = mode(nnhood_label, axis=1, keepdims=False)
                     nnhood_bool = np.ravel(nnhood_label) == y_class
                 elif self.kind_sel == "all":
                     nnhood_label = nnhood_label == target_class
